refactor: create useConfig custom hook
src/pages/Start.jsx에서 기존에 하드 코딩 되어있던 configState와 관련 함수들을 하나로 묶어 useConfig hook으로 재정의.
